исправлен дурацкий баг, который мешал пересоздавать объект парсера провел дебагом тесты переполнения стека и изучил target на предмет наличия всякого шлака

+++ Downloads/SoftWerkeProject/src/main/java/com/nikolay/imbirev/connector/queryform/Performer.java
@@ -21,7 +21,10 @@ class Performer {
                     return service.updateTable(searhQueries, insertOrUpdateQueries);
                 } else return RequestCode.EMPTY_SET;
             case "delete":
-                return service.deleteFromTable(insertOrUpdateQueries);
+                RequestCode code1 = service.getFromTable(searhQueries, new Column[]{});
+                if (code1 == RequestCode.SUCCESS) {
+                    return service.deleteFromTable(insertOrUpdateQueries);
+                } else return RequestCode.EMPTY_SET;
             case "get":
                 return service.getFromTable(searhQueries, sortColumns);
             default:
